requirements.txt
# Основные зависимости для парсера Kwork.ru
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# База данных
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Telegram бот
python-telegram-bot==20.7

# Конфигурация и валидация
pydantic>=1.10.0,<3.0.0
PyYAML==6.0.1

# Экспорт данных
pandas==2.1.4
openpyxl==3.1.2

# Утилиты
asyncio-throttle==1.0.2
fake-useragent==1.4.0
tenacity==8.2.3

# Логирование и мониторинг
loguru==0.7.2

# CLI
click==8.1.7
rich==13.7.0

# Развертывание
uvloop==0.19.0; sys_platform != "win32"

# Тестирование (опционально)
pytest==7.4.3
pytest-asyncio==0.21.1 